move TestIsZeroHashP2PHKAddress to txthelpers_test.go
